Шаблонизация исправления
