+++ data.php
@@ -10,47 +10,43 @@ $posters = [
     'name' => '2014 Rossignol District Snowboard',
     'categories' => 'Доски и лыжи',
     'price' => '10999',
-    'URL' => 'img/lot-1.jpg'
+    'URL' => 'img/lot-1.jpg',
+    'date_end' => "2022-04-08"
     ],
     [
     'name' => 'DC Ply Mens 2016/2017 Snowboard',
     'categories' => 'Доски и лыжи',
     'price' => '159999',
-    'URL' => 'img/lot-2.jpg'
+    'URL' => 'img/lot-2.jpg',
+    'date_end' => "2022-04-08"
     ],
     [
     'name' => 'Крепления Union Contact Pro 2015 года размер L/XL',
     'categories' => 'Крепления',
     'price' => '8000',
-    'URL' => 'img/lot-3.jpg'
+    'URL' => 'img/lot-3.jpg',
+    'date_end' => "2022-04-08"
     ],
     [
     'name' => 'Ботинки для сноуборда DC Mutiny Charocal',
     'categories' => 'Ботинки',
     'price' => '10999',
-    'URL' => 'img/lot-4.jpg'
+    'URL' => 'img/lot-4.jpg',
+    'date_end' => "2022-04-08"
     ],
     [
     'name' => 'Куртка для сноуборда DC Mutiny Charocal',
     'categories' => 'Одежда',
     'price' => '7500',
-    'URL' => 'img/lot-5.jpg'
+    'URL' => 'img/lot-5.jpg',
+    'date_end' => "2022-04-08"
     ],
     [
     'name' => 'Маска Oakley Canopy',
     'categories' => 'Разное',
     'price' => '5400',
-    'URL' => 'img/lot-6.jpg'
+    'URL' => 'img/lot-6.jpg',
+    'date_end' => "2022-04-08"
     ]
 ];
-
-function priceModify(int $price): string {
-    ceil($price);
-    if ($price > 1000) {
-        $price = number_format($price, 0, '', ' ');
-    }
-    return (string) $price . ' ₽';
-}
-
-
 ?>
