+++ functions.php
@@ -1,0 +1,11 @@
+<?php
+
+function priceModify(int $price): string {
+    ceil($price);
+    if ($price > 1000) {
+        $price = number_format($price, 0, '', ' ');
+    }
+    return (string) $price . ' ₽';
+}
+
+?>
